requirements.txt
pymongo==3.11.3
Flask==1.1.2
Flask_Cors==3.0.10
Flask_PyMongo==2.3.0
dnspython==2.1.0
